Convenience functions

Convenience functions

ERM implements the following utility functions.

EmpiricalRiskMinimization.matrix — Method.

Convert an n-vector to an nx1 array

source

EmpiricalRiskMinimization.sigm — Method.

Sigmoid function

source

EmpiricalRiskMinimization.findvalue — Method.

findvalue(s, lst): Find the first index of s in a list lst

source

EmpiricalRiskMinimization.rms — Method.

compute the rms of a matrix or a vector

source

EmpiricalRiskMinimization.rowwise — Method.

Apply f to each row of a matrix. f should map vectors to vectors

source