Losses

Losses

These are all the available loss functions exported by ERM. See the corresponding usage page to understand how and when to use these functions.

EmpiricalRiskMinimization.AbsoluteLoss — Type.

AbsoluteLoss() constructs the l1/absolute loss. Use with Model().

source

EmpiricalRiskMinimization.HingeLoss — Type.

HingeLoss() constructs the hinge loss (i.e., for SVM). Use withModel()`.

source

EmpiricalRiskMinimization.HuberLoss — Type.

HuberLoss() constructs Huber loss. Use with Model(). Can also invoke as HuberLoss(alpha), which allows specification of the tradeoff parameter alpha > 0. Note that HuberLoss() defaults to alpha = 1.0.

source

EmpiricalRiskMinimization.LogisticLoss — Type.

LogisticLoss() constructs the logistic loss for classification problems. Use with Model()

source

EmpiricalRiskMinimization.SigmoidLoss — Type.

SigmoidLoss() constructs the sigmoid loss for classification problems. Use with Model()

source

EmpiricalRiskMinimization.SquareLoss — Type.

SquareLoss() constructs the l2/squared loss. Use with Model().

source