Models and training

Models and training

EmpiricalRiskMinimization.Model — Type.

Model(...)

The Model() function constructs an ERM model. The typical invocation is Model(U, V, Loss(), Reg()), where U and V specify raw inputs and targets, respectively, and Loss() specifies some type of training loss (default: SquareLoss()) and Reg() specifies some type of regularizer (default: L2Reg(1.0)). For more details, see the description of ERM models in the usage notes.

source

EmpiricalRiskMinimization.predict_v_from_test — Function.

predict_v_from_test(M [, theta])

Allows you compute unembedded predictions (i.e., in V space) based on a trained ERM model M on test data. Option to specify a choice of theta. It defaults to theta=thetaopt(M)

source

EmpiricalRiskMinimization.predict_v_from_train — Function.

predict_v_from_train(M [, theta])

Allows you compute unembedded predictions (i.e., in V space) based on a trained ERM model M on train data. Option to specify a choice of theta. It defaults to theta=thetaopt(M)

source

EmpiricalRiskMinimization.predict_v_from_u — Method.

predict_y_from_u(M, U [, theta])

Allows you compute unembedded predictions (i.e., in V space) based on a trained ERM model M on one or many raw inputs, U. Option to specify a choice of theta. It defaults to theta=thetaopt(M)

source

EmpiricalRiskMinimization.predict_y_from_test — Function.

predict_y_from_test(M [, theta])

Allows you compute embedded predictions (i.e., y values) based on a trained ERM model M on test data. Option to specify a choice of theta. It defaults to theta=thetaopt(M)

source

EmpiricalRiskMinimization.predict_y_from_train — Function.

predict_y_from_train(M [, theta])

Allows you compute embedded predictions (i.e., y values) based on a trained ERM model M on train data. Option to specify a choice of theta. It defaults to theta=thetaopt(M)

source

EmpiricalRiskMinimization.predict_y_from_u — Method.

predict_y_from_u(M, U [, theta])

Allows you compute embedded predictions (i.e., y values) based on a trained ERM model M on one or many raw inputs, U. Option to specify a choice of theta. It defaults to theta=thetaopt(M)

source

EmpiricalRiskMinimization.status — Method.

Prints and returns the status of the model.

source