Prediction

Prediction

The functions listed below are all the methods made available by ERM for predicting y and v values on training and testing data. See the corresponding usage page to understand how to use these functions.

EmpiricalRiskMinimization.predict_v_from_test — Function.

predict_v_from_test(M [, theta])

Allows you compute unembedded predictions (i.e., in V space) based on a trained ERM model M on test data. Option to specify a choice of theta. It defaults to theta=thetaopt(M)

source

EmpiricalRiskMinimization.predict_v_from_train — Function.

predict_v_from_train(M [, theta])

Allows you compute unembedded predictions (i.e., in V space) based on a trained ERM model M on train data. Option to specify a choice of theta. It defaults to theta=thetaopt(M)

source

EmpiricalRiskMinimization.predict_v_from_u — Method.

predict_y_from_u(M, U [, theta])

Allows you compute unembedded predictions (i.e., in V space) based on a trained ERM model M on one or many raw inputs, U. Option to specify a choice of theta. It defaults to theta=thetaopt(M)

source

EmpiricalRiskMinimization.predict_y_from_test — Function.

predict_y_from_test(M [, theta])

Allows you compute embedded predictions (i.e., y values) based on a trained ERM model M on test data. Option to specify a choice of theta. It defaults to theta=thetaopt(M)

source

EmpiricalRiskMinimization.predict_y_from_train — Function.

predict_y_from_train(M [, theta])

Allows you compute embedded predictions (i.e., y values) based on a trained ERM model M on train data. Option to specify a choice of theta. It defaults to theta=thetaopt(M)

source

EmpiricalRiskMinimization.predict_y_from_u — Method.

predict_y_from_u(M, U [, theta])

Allows you compute embedded predictions (i.e., y values) based on a trained ERM model M on one or many raw inputs, U. Option to specify a choice of theta. It defaults to theta=thetaopt(M)

source

EmpiricalRiskMinimization.confusion_test — Method.

confusion_test(M)

Returns the confusion matrix based on test data for a trained ERM model

source

EmpiricalRiskMinimization.confusion_train — Method.

confusion_train(M)

Returns the confusion matrix based on training data for a trained ERM model

source