Validation and out-of-sample testing

Validation and out-of-sample testing

These are all the available library functions that users have available for validation and out-of-sample testing. Refer to the corresponding usage page to see examples and explanations behind these functions.

EmpiricalRiskMinimization.lambda — Method.

lambda(M) returns the lambda used during training. If multiple lambda values were specified, returns the optimal one

source

EmpiricalRiskMinimization.lambdaopt — Method.

lambdaopt(M) returns the optimal lambda computed during training of M

source

EmpiricalRiskMinimization.lambdapath — Method.

lambdapath(M) returns the vector of lambdas used

source

EmpiricalRiskMinimization.testloss — Method.

testloss(M) returns the average testing loss

source

EmpiricalRiskMinimization.testlosspath — Method.

testlosspath(M) returns the vector of test losses at various lambdas

source

EmpiricalRiskMinimization.thetaopt — Method.

thetaopt(M) returns the optimal theta chosen during training

source

EmpiricalRiskMinimization.thetapath — Method.

thetapath(M) returns the vector of optimal thetas at various lambdas

source

EmpiricalRiskMinimization.trainloss — Method.

trainloss(M) returns the average training loss

source

EmpiricalRiskMinimization.trainlosspath — Method.

trainlosspath(M) returns the vector of train losses at various lambdas

source